requirements.txt
tqdm
python-dotenv
transformers
datasets
langchain
langchain-openai
langchain-huggingface